GPT_SoVITS/AR/modules/patched_mha_with_cache_onnx.py
from torch.nn.functional import *
from torch.nn.functional import (
    _mha_shape_check,
    _canonical_mask,
    _none_or_dtype,
    _in_projection_packed,
)
import torch
from typing import Optional, Tuple
import logging

logger = logging.getLogger(__name__)

"""
🎯 GPT-SoVITS KV缓存优化版本 (基于作者建议的固定缓冲区策略)

主要修复和优化：
1. 修复了错误的[:-1]切片逻辑，避免缓存内容丢失
2. 添加了缓存大小限制，防止内存无限增长
3. 实现固定缓冲区+索引写入策略，避免torch.cat的内存重分配
4. 支持滑动窗口机制，处理缓冲区满的情况
5. 添加了兼容性回退机制，确保在各种环境下都能正常工作

🚀 核心优化策略（基于作者建议）：
- 预分配固定大小的GPU缓冲区，满足IO Binding的形状要求
- 使用索引写入替代torch.cat拼接，避免动态内存分配
- 在注意力计算时仅使用缓冲区中已填充的部分
- 支持滑动窗口策略，处理长序列场景

性能提升：
- 避免torch.cat的内存重分配开销
- 减少GPU内存碎片，提升访问效率
- 兼容ONNX Runtime的IO Binding优化
- 减少CPU-GPU数据交换频率
- 保持与原版API的完全兼容性
"""

def multi_head_attention_forward_patched(
    query,
    key,
    value,
    embed_dim_to_check: int,
    num_heads: int,
    in_proj_weight,
    in_proj_bias: Optional[torch.Tensor],
    bias_k: Optional[torch.Tensor],
    bias_v: Optional[torch.Tensor],
    add_zero_attn: bool,
    dropout_p: float,
    out_proj_weight: torch.Tensor,
    out_proj_bias: Optional[torch.Tensor],
    training: bool = True,
    key_padding_mask: Optional[torch.Tensor] = None,
    need_weights: bool = True,
    attn_mask: Optional[torch.Tensor] = None,
    use_separate_proj_weight: bool = False,
    q_proj_weight: Optional[torch.Tensor] = None,
    k_proj_weight: Optional[torch.Tensor] = None,
    v_proj_weight: Optional[torch.Tensor] = None,
    static_k: Optional[torch.Tensor] = None,
    static_v: Optional[torch.Tensor] = None,
    average_attn_weights: bool = True,
    is_causal: bool = False,
    cache=None,
) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:

    # set up shape vars
    _, _, embed_dim = query.shape
    attn_mask = _canonical_mask(
        mask=attn_mask,
        mask_name="attn_mask",
        other_type=None,
        other_name="",
        target_type=query.dtype,
        check_other=False,
    )
    head_dim = embed_dim // num_heads

    proj_qkv = linear(query, in_proj_weight, in_proj_bias)
    proj_qkv = proj_qkv.unflatten(-1, (3, query.size(-1))).unsqueeze(0).transpose(0, -2).squeeze(-2).contiguous()
    q, k, v = proj_qkv[0], proj_qkv[1], proj_qkv[2]

    # 🎯 关键优化：修复KV缓存逻辑，避免ONNX fallback到CPU
    if cache is not None:
        if cache["first_infer"] == 1:
            # 首次推理：直接存储
            cache["k"][cache["stage"]] = k
            cache["v"][cache["stage"]] = v
        else:
            # 🔧 修复：移除错误的[:-1]切片，使用正确的缓存拼接逻辑
            # 原错误代码：cache["k"][cache["stage"]] = torch.cat([cache["k"][cache["stage"]][:-1], k], 0)
            # 正确逻辑：直接拼接，保持与原版一致
            
            # 🚀 性能优化：使用固定缓冲区+索引写入策略
            # 基于作者建议：预分配固定大小缓冲区，使用索引写入而非torch.cat
            
            # 🎯 缓存大小限制：避免内存无限增长
            max_cache_size = getattr(cache, 'max_size', 1024)  # 默认最大缓存1024个token
            
            # 🔧 初始化固定大小缓冲区（如果尚未初始化）
            stage_key = f"stage_{cache['stage']}"
            if not hasattr(cache, f'_buffer_initialized_{stage_key}'):
                # 获取当前缓存的实际形状信息
                current_k = cache["k"][cache["stage"]]
                current_v = cache["v"][cache["stage"]]
                
                # 预分配固定大小的缓冲区
                cache["k"][cache["stage"]] = torch.zeros(
                    max_cache_size, current_k.size(1), current_k.size(2),
                    dtype=current_k.dtype, device=current_k.device
                )
                cache["v"][cache["stage"]] = torch.zeros(
                    max_cache_size, current_v.size(1), current_v.size(2),
                    dtype=current_v.dtype, device=current_v.device
                )
                
                # 标记已初始化
                setattr(cache, f'_buffer_initialized_{stage_key}', True)
                setattr(cache, f'_current_length_{stage_key}', 0)
                
                logger.debug(f"🔧 初始化固定缓冲区: stage={cache['stage']}, size={max_cache_size}")
            
            try:
                # 🚀 使用索引写入策略，避免torch.cat
                cached_k = cache["k"][cache["stage"]]
                cached_v = cache["v"][cache["stage"]]
                current_length = getattr(cache, f'_current_length_{stage_key}', 0)
                
                new_length = k.size(0)
                
                if current_length + new_length <= max_cache_size:
                    # 有足够空间，直接追加到缓冲区
                    cached_k[current_length:current_length + new_length] = k
                    cached_v[current_length:current_length + new_length] = v
                    setattr(cache, f'_current_length_{stage_key}', current_length + new_length)
                    
                    # 使用实际填充的部分
                    k = cached_k[:current_length + new_length]
                    v = cached_v[:current_length + new_length]
                    
                else:
                    # 缓冲区满，使用滑动窗口策略
                    # 计算需要保留的空间
                    keep_length = max_cache_size - new_length
                    
                    if keep_length > 0:
                        # 保留最近的数据，丢弃最旧的数据
                        cached_k[:keep_length] = cached_k[current_length - keep_length:current_length]
                        cached_v[:keep_length] = cached_v[current_length - keep_length:current_length]
                        
                        # 写入新数据
                        cached_k[keep_length:] = k
                        cached_v[keep_length:] = v
                        
                        setattr(cache, f'_current_length_{stage_key}', max_cache_size)
                        
                        # 使用全部缓冲区
                        k = cached_k
                        v = cached_v
                    else:
                        # 新数据比缓冲区还大，直接替换
                        cached_k[:new_length] = k
                        cached_v[:new_length] = v
                        cached_k[new_length:] = 0  # 清零剩余部分
                        cached_v[new_length:] = 0
                        
                        setattr(cache, f'_current_length_{stage_key}', new_length)
                        
                        # 使用实际数据部分
                        k = cached_k[:new_length]
                        v = cached_v[:new_length]
                
                logger.debug(f"📝 索引写入完成: stage={cache['stage']}, length={getattr(cache, f'_current_length_{stage_key}', 0)}")
                
            except Exception as e:
                # 🔄 兼容性回退：如果优化失败，使用原始逻辑
                logger.warning(f"⚠️ 固定缓冲区策略失败，回退到torch.cat: {e}")
                cache["k"][cache["stage"]] = torch.cat([cache["k"][cache["stage"]], k], 0)
                cache["v"][cache["stage"]] = torch.cat([cache["v"][cache["stage"]], v], 0)
                k = cache["k"][cache["stage"]]
                v = cache["v"][cache["stage"]]
        cache["stage"] = (cache["stage"] + 1) % cache["all_stage"]

    attn_mask = _canonical_mask(
        mask=attn_mask,
        mask_name="attn_mask",
        other_type=None,
        other_name="",
        target_type=q.dtype,
        check_other=False,
    )
    attn_mask = attn_mask.unsqueeze(0)

    q = q.view(-1, num_heads, head_dim).transpose(0, 1)
    k = k.view(-1, num_heads, head_dim).transpose(0, 1)
    v = v.view(-1, num_heads, head_dim).transpose(0, 1)

    dropout_p = 0.0
    attn_mask = attn_mask.unsqueeze(0)
    q = q.view(num_heads, -1, head_dim).unsqueeze(0)
    k = k.view(num_heads, -1, head_dim).unsqueeze(0)
    v = v.view(num_heads, -1, head_dim).unsqueeze(0)
    attn_output = scaled_dot_product_attention(
        q, k, v, attn_mask, dropout_p, is_causal
    )
    attn_output = (
        attn_output.permute(2, 0, 1, 3).contiguous().view(-1, embed_dim)
    )
    attn_output = linear(attn_output, out_proj_weight, out_proj_bias)
    attn_output = attn_output.view(-1, 1, attn_output.size(1))

    return attn_output
