requirements.txt
# =============================================================================
# amadeus-realtime — Python Dependencies / Python 依赖
#
# Installation order / 安装顺序：
#   1. Install PyTorch for your CUDA version first / 先按 CUDA 版本安装 PyTorch：
#      https://pytorch.org/get-started/locally/
#   2. pip install -r requirements.txt
# =============================================================================

# ── GPT-SoVITS Core Inference / GPT-SoVITS 核心推理 ──────────────────────────
numpy==1.23.4
scipy
librosa==0.9.2
numba==0.56.4
torchaudio
soundfile
# torch: install separately per CUDA version / 需单独按 CUDA 版本安装

# TTS linguistic processing / TTS 语言处理
cn2an
pypinyin
pyopenjtalk>=0.3.4
g2p_en
sentencepiece
transformers>=4.43
tokenizers>=0.13,<1
tqdm
jieba
jieba_fast
split-lang
fast_langdetect>=0.3.0
wordsegment
rotary_embedding_torch
ToJyutping
g2pk2
ko_pron
opencc; sys_platform != 'linux'
opencc==1.1.1; sys_platform == 'linux'

# TTS post-processing / TTS 后处理
x_transformers
torchmetrics<=1.5
av>=11
onnxruntime; sys_platform == 'darwin'
onnxruntime-gpu; sys_platform != 'darwin'

# ── LLM Clients / LLM 客户端 ─────────────────────────────────────────────────
# DeepSeek / LM Studio / llama-server (OpenAI-compatible) / 兼容 OpenAI 接口
openai>=1.0.0
# Gemini Flash — used by stream_llm_query / 用于 stream_llm_query 流式请求
google-generativeai>=0.8.0
# Google Live API — used by live_api_sidecar.py / 实时双向语音对话
google-genai>=1.0.0
# Connection pooling for LLM calls / LLM 调用连接池
httpx>=0.24.0
# AWS Bedrock (optional) / AWS Bedrock 支持（可选）
boto3>=1.28.0
requests>=2.28.0
aiohttp>=3.8.0

# ── GUI ──────────────────────────────────────────────────────────────────────
PyQt5>=5.15.0
# Qt + asyncio event loop bridge / Qt 与 asyncio 事件循环桥接
qasync>=0.23.0

# ── Audio I/O / 语音输入输出 ──────────────────────────────────────────────────
# Microphone recording & audio playback / 麦克风录音与音频输出
pyaudio>=0.2.13
# ASR abstraction layer / 语音识别封装层
SpeechRecognition>=3.10.0
# Local Whisper ASR accelerated backend / 本地 Whisper ASR 加速后端
faster-whisper>=1.0.0

# ── VTuber Studio / VTuber Studio 连接 ───────────────────────────────────────
# WebSocket communication with VTS / 与 VTS 通信
websocket-client>=1.6.0

# ── Multimodal Input / 多模态输入 ────────────────────────────────────────────
# High-performance screen capture / 高性能屏幕截图
mss>=9.0.0
# Image processing for screenshots / 截图图像处理
Pillow>=10.0.0

# ── RAG Knowledge Base / RAG 知识库 ──────────────────────────────────────────
# Vector similarity search / 向量相似度检索
faiss-cpu>=1.7.4           # GPU machines: replace with faiss-gpu / GPU 机器可换 faiss-gpu
# Text embedding for RAG / RAG 文本嵌入
sentence-transformers>=2.2.0

# ── Configuration / 配置管理 ─────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Misc / 其他工具 ───────────────────────────────────────────────────────────
sympy>=1.12                # Unit conversion (internal) / 单位转换（内部使用）
chardet
PyYAML
psutil
# LoRA weight loading for inference / 推理侧 LoRA 权重加载
peft
huggingface_hub>=0.13
